requirements.txt
duckdb==1.4.3
pandas==2.3.3
pyarrow==23.0.0
openpyxl==3.1.5
rapidfuzz==3.14.1
pytest==8.4.1
google-genai
python-dotenv

